requirements.txt
python-telegram-bot==20.7
Flask==3.0.2
requests==2.31.0
